% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_endpoints}
\alias{api_endpoints}
\title{The API endpoints at the Danish Parliament}
\format{A vector}
\source{
\url{http://www.ft.dk/dokumenter/aabne_data/}
}
\usage{
api_endpoints
}
\description{
A vector containing the endpoints for the API.
}
\keyword{datasets}
