% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{null_to_na}
\alias{null_to_na}
\title{Turns NULL values in a list into NAs.}
\usage{
null_to_na(mylist)
}
\arguments{
\item{mylist}{is a list, where the NULL values are to be turned into NAs.}
}
\description{
Turns NULL values in a list into NAs.
}
