% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{see_endpoints}
\alias{see_endpoints}
\title{See the available endpoints in the API}
\usage{
see_endpoints()
}
\value{
A vector of possible endpoints
}
\description{
See the available endpoints in the API
}
\examples{
library(ftDK)
see_endpoints()
}
