% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ft}
\alias{get_ft}
\title{Get data from the API}
\usage{
get_ft(endpoint)
}
\arguments{
\item{endpoint}{is the endpoint in the API that you want data from.}
}
\value{
a data frame with the data requested
}
\description{
Get data from the API
}
\examples{
library(ftDK)
see_endpoints()

ft_data  <- get_ft("afstemning")
ft_data
}
